+++ traffic_ops/traffic_ops_golang/deliveryservice/keys.go
@@ -42,7 +42,8 @@ import (
 )
 
 const (
-	PemCertEndMarker = "-----END CERTIFICATE-----"
+	PemCertEndMarker  = "-----END CERTIFICATE-----"
+	hostnameKeyDepMsg = "This endpoint is deprecated, please use '/deliveryservices/xmlId/{{XMLID}}/sslkeys' instead"
 )
 
 // AddSSLKeys adds the given ssl keys to the given delivery service.
@@ -130,49 +131,53 @@ func AddSSLKeys(w http.ResponseWriter, r *http.Request) {
 
 // GetSSLKeysByHostName fetches the ssl keys for a deliveryservice specified by the fully qualified hostname
 func GetSSLKeysByHostName(w http.ResponseWriter, r *http.Request) {
-	inf, userErr, sysErr, errCode := api.NewInfo(r, []string{"hostname"}, nil)
-	if userErr != nil || sysErr != nil {
-		api.HandleErr(w, r, inf.Tx.Tx, errCode, userErr, sysErr)
+	inf, xmlID, err := getXmlIDFromRequest(w, r)
+	if inf != nil {
+		defer inf.Close()
+	}
+	if err != nil {
 		return
 	}
-	defer inf.Close()
+	getSSLKeysByXMLIDHelper(xmlID, inf, w, r)
+}
+
+func getXmlIDFromRequest(w http.ResponseWriter, r *http.Request) (*api.APIInfo, string, error) {
+	alerts := tc.CreateAlerts(tc.WarnLevel, hostnameKeyDepMsg)
+	inf, userErr, sysErr, errCode := api.NewInfo(r, []string{"hostname"}, nil)
+	if userErr != nil || sysErr != nil {
+		userErr = api.LogErr(r, errCode, userErr, sysErr)
+		alerts.AddNewAlert(tc.ErrorLevel, userErr.Error())
+		api.WriteAlerts(w, r, errCode, alerts)
+		return inf, "", errors.New("getting XML ID from request")
+	}
 
 	if inf.Config.RiakEnabled == false {
-		api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, nil, errors.New("getting SSL keys from Riak by host name: Riak is not configured"))
-		return
+		userErr = api.LogErr(r, http.StatusInternalServerError, nil, errors.New("getting SSL keys from Riak by host name: Riak is not configured"))
+		alerts.AddNewAlert(tc.ErrorLevel, userErr.Error())
+		api.WriteAlerts(w, r, http.StatusInternalServerError, alerts)
+		return inf, "", errors.New("getting XML ID from request")
 	}
 
 	hostName := inf.Params["hostname"]
 	xmlID, userErr, sysErr, errCode := getXmlIdFromHostname(inf, hostName)
 	if userErr != nil || sysErr != nil {
-		api.HandleErr(w, r, inf.Tx.Tx, errCode, userErr, sysErr)
-		return
+		userErr = api.LogErr(r, errCode, userErr, sysErr)
+		alerts.AddNewAlert(tc.ErrorLevel, userErr.Error())
+		api.WriteAlerts(w, r, errCode, alerts)
+		return inf, "", errors.New("getting XML ID from request")
 	}
-
-	getSSLKeysByXMLIDHelper(xmlID, inf, w, r)
+	return inf, xmlID, nil
 }
 
 // GetSSLKeysByHostNameV15 fetches the ssl keys for a deliveryservice specified by the fully qualified hostname. V15 includes expiration date.
 func GetSSLKeysByHostNameV15(w http.ResponseWriter, r *http.Request) {
-	inf, userErr, sysErr, errCode := api.NewInfo(r, []string{"hostname"}, nil)
-	if userErr != nil || sysErr != nil {
-		api.HandleErr(w, r, inf.Tx.Tx, errCode, userErr, sysErr)
+	inf, xmlID, err := getXmlIDFromRequest(w, r)
+	if inf != nil {
+		defer inf.Close()
+	}
+	if err != nil {
 		return
 	}
-	defer inf.Close()
-
-	if inf.Config.RiakEnabled == false {
-		api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, nil, errors.New("getting SSL keys from Riak by host name: Riak is not configured"))
-		return
-	}
-
-	hostName := inf.Params["hostname"]
-	xmlID, userErr, sysErr, errCode := getXmlIdFromHostname(inf, hostName)
-	if userErr != nil || sysErr != nil {
-		api.HandleErr(w, r, inf.Tx.Tx, errCode, userErr, sysErr)
-		return
-	}
-
 	getSSLKeysByXMLIDHelperV15(xmlID, inf, w, r)
 }
 
@@ -226,30 +231,38 @@ func GetSSLKeysByXMLID(w http.ResponseWriter, r *http.Request) {
 }
 
 func getSSLKeysByXMLIDHelper(xmlID string, inf *api.APIInfo, w http.ResponseWriter, r *http.Request) {
+	alerts := tc.CreateAlerts(tc.WarnLevel, hostnameKeyDepMsg)
 	version := inf.Params["version"]
 	decode := inf.Params["decode"]
 	if userErr, sysErr, errCode := tenant.Check(inf.User, xmlID, inf.Tx.Tx); userErr != nil || sysErr != nil {
-		api.HandleErr(w, r, inf.Tx.Tx, errCode, userErr, sysErr)
+		userErr = api.LogErr(r, errCode, userErr, sysErr)
+		alerts.AddNewAlert(tc.ErrorLevel, userErr.Error())
+		api.WriteAlerts(w, r, errCode, alerts)
 		return
 	}
 	keyObj, ok, err := riaksvc.GetDeliveryServiceSSLKeysObj(xmlID, version, inf.Tx.Tx, inf.Config.RiakAuthOptions, inf.Config.RiakPort)
 	if err != nil {
-		api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, nil, errors.New("getting ssl keys: "+err.Error()))
+		userErr := api.LogErr(r, http.StatusInternalServerError, nil, errors.New("getting ssl keys: "+err.Error()))
+		alerts.AddNewAlert(tc.ErrorLevel, userErr.Error())
+		api.WriteAlerts(w, r, http.StatusInternalServerError, alerts)
 		return
 	}
 	if !ok {
-		api.WriteAlertsObj(w, r, http.StatusNotFound, tc.CreateAlerts(tc.ErrorLevel, "no object found for the specified key"), struct{}{}) // empty response object because Perl
+		alerts.AddNewAlert(tc.ErrorLevel, "no object found for the specified key")
+		api.WriteAlertsObj(w, r, http.StatusNotFound, alerts, struct{}{}) // empty response object because Perl
 		return
 	}
 	if decode != "" && decode != "0" { // the Perl version checked the decode string as: if ( $decode )
 		err = base64DecodeCertificate(&keyObj.Certificate)
 		if err != nil {
-			api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, nil, errors.New("getting SSL keys for XMLID '"+xmlID+"': "+err.Error()))
+			userErr := api.LogErr(r, http.StatusInternalServerError, nil, errors.New("getting SSL keys for XMLID '"+xmlID+"': "+err.Error()))
+			alerts.AddNewAlert(tc.ErrorLevel, userErr.Error())
+			api.WriteAlerts(w, r, http.StatusInternalServerError, alerts)
 			return
 		}
 	}
 
-	api.WriteResp(w, r, keyObj)
+	api.WriteAlertsObj(w, r, http.StatusOK, alerts, keyObj)
 }
 
 // GetSSLKeysByXMLIDV15 fetches the deliveryservice ssl keys by the specified xmlID. V15 includes expiration date.
@@ -269,26 +282,34 @@ func GetSSLKeysByXMLIDV15(w http.ResponseWriter, r *http.Request) {
 }
 
 func getSSLKeysByXMLIDHelperV15(xmlID string, inf *api.APIInfo, w http.ResponseWriter, r *http.Request) {
+	alerts := tc.CreateAlerts(tc.WarnLevel, hostnameKeyDepMsg)
 	version := inf.Params["version"]
 	decode := inf.Params["decode"]
 	if userErr, sysErr, errCode := tenant.Check(inf.User, xmlID, inf.Tx.Tx); userErr != nil || sysErr != nil {
-		api.HandleErr(w, r, inf.Tx.Tx, errCode, userErr, sysErr)
+		userErr = api.LogErr(r, errCode, userErr, sysErr)
+		alerts.AddNewAlert(tc.ErrorLevel, userErr.Error())
+		api.WriteAlerts(w, r, errCode, alerts)
 		return
 	}
 	keyObj, ok, err := riaksvc.GetDeliveryServiceSSLKeysObjV15(xmlID, version, inf.Tx.Tx, inf.Config.RiakAuthOptions, inf.Config.RiakPort)
 	if err != nil {
-		api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, nil, errors.New("getting ssl keys: "+err.Error()))
+		userErr := api.LogErr(r, http.StatusInternalServerError, nil, errors.New("getting ssl keys: "+err.Error()))
+		alerts.AddNewAlert(tc.ErrorLevel, userErr.Error())
+		api.WriteAlerts(w, r, http.StatusInternalServerError, alerts)
 		return
 	}
 	if !ok {
-		api.WriteAlertsObj(w, r, http.StatusNotFound, tc.CreateAlerts(tc.ErrorLevel, "no object found for the specified key"), struct{}{}) // empty response object because Perl
+		alerts.AddNewAlert(tc.ErrorLevel, "no object found for the specified key")
+		api.WriteAlertsObj(w, r, http.StatusNotFound, alerts, struct{}{}) // empty response object because Perl
 		return
 	}
 
 	parsedCert := keyObj.Certificate
 	err = base64DecodeCertificate(&parsedCert)
 	if err != nil {
-		api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, nil, errors.New("getting SSL keys for XMLID '"+xmlID+"': "+err.Error()))
+		userErr := api.LogErr(r, http.StatusInternalServerError, nil, errors.New("getting SSL keys for XMLID '"+xmlID+"': "+err.Error()))
+		alerts.AddNewAlert(tc.ErrorLevel, userErr.Error())
+		api.WriteAlerts(w, r, http.StatusInternalServerError, alerts)
 		return
 	}
 	if decode != "" && decode != "0" { // the Perl version checked the decode string as: if ( $decode )
@@ -298,12 +319,15 @@ func getSSLKeysByXMLIDHelperV15(xmlID string, inf *api.APIInfo, w http.ResponseW
 	if keyObj.Certificate.Crt != "" && keyObj.Expiration.IsZero() {
 		exp, err := parseExpirationFromCert([]byte(parsedCert.Crt))
 		if err != nil {
-			api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, nil, errors.New(xmlID+": "+err.Error()))
+			userErr := api.LogErr(r, http.StatusInternalServerError, nil, errors.New(xmlID+": "+err.Error()))
+			alerts.AddNewAlert(tc.ErrorLevel, userErr.Error())
+			api.WriteAlerts(w, r, http.StatusInternalServerError, alerts)
 			return
 		}
 		keyObj.Expiration = exp
 	}
-	api.WriteResp(w, r, keyObj)
+
+	api.WriteAlertsObj(w, r, http.StatusOK, alerts, keyObj)
 }
 
 func parseExpirationFromCert(cert []byte) (time.Time, error) {