+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - /cdns/usage/overview
   - /deliveryservice_user
   - /deliveryservice_user/:dsId/:userId
+  - /deliveryservices/hostname/:name/sslkeys
   - /deliveryservices/:id/state
   - /divisions/:division_name/regions
   - /divisions/name/:name